+++ front-end/src/App.tsx
@@ -14,6 +14,7 @@ import { UserDetailsProvider } from './context/UserDetailsContext';
 import CreatePost from './screens/CreatePost';
 import Discussions from './screens/Discussions';
 import Home from './screens/Home';
+import { PrivacyPolicy, TermsAndConditions, TermsOfWebsite } from './screens/LegalDocuments';
 import LoginForm from './screens/LoginForm';
 import NotFound from './screens/NotFound';
 import PostDiscussion from './screens/DiscussionPost';
@@ -21,13 +22,10 @@ import OnChain from './screens/OnChain';
 import PostMotion from './screens/MotionPost';
 import PostProposal from './screens/ProposalPost';
 import PostReferendum from './screens/ReferendumPost';
-import PrivacyPolicy from './screens/PrivacyPolicy';
 import RequestResetPassword from './screens/RequestResetPassword';
 import ResetPassword from './screens/RequestResetPassword/ResetPassword';
 import Settings from './screens/Settings';
 import SignupForm from './screens/SignupForm';
-import TermsAndConditions from './screens/TermsAndConditions';
-import TermsOfWebsite from './screens/TermsOfWebsite';
 import VerifyEmail from './screens/VerifyEmail';
 import UndoEmailChange from './screens/UndoEmailChange';
 import { theme } from './themes/theme';
